Move test_llama_lora to integration tests
