+++ src/fairseq2/nn/lora.py
@@ -1,3 +1,9 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
 import re
 from dataclasses import dataclass
 from typing import Any, Dict, List, Literal, Optional, final
